+++ CHANGES.md
@@ -5,6 +5,8 @@ CHANGES - 2.3b1 - 2017-11-01
 Changes in CUPS v2.3b1
 ----------------------
 
+- No longer support backslash, question mark, or quotes in printer names
+  (Issue #4966)
 - Dropped RSS subscription management from the web interface (Issue #5012)
 - The lpadmin command now provides a better error message when an unsupported
   System V interface script is used (Issue #5111)